Add files via upload
This program takes a .txt file and parses the data, and outputs it into a .txt file at the same location. I have included the input file, the main.cpp file and the resulting output file. The location of the input and output files is given to the program by the IDE and parsed so the program can find it.  

Adam Claxton